requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
cffi==1.11.2
cryptography==2.1.1
Django==1.11.6
django-webpack-loader==0.5.0
djangorestframework==3.7.0
Fabric==1.14.0
idna==2.6
paramiko==2.3.1
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.1.2
pytz==2017.2
six==1.11.0